audio-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-socketio==5.10.0
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
librosa==0.10.1
numpy==1.26.3
scipy==1.12.0
yt-dlp==2024.1.0
aiofiles==23.2.1
